Moving selector specific tests to their own testcase
